allow not creating/managing the S3 bucket (#6)
* allow not creating/managing the S3 bucket

This allows to have multiple CloudFront distributions pointing to the
same S3 bucket, such as to different paths.

Note, due to the way the conditional parser works in current version of
Terraform we're using the workaround using join() and the .*. operator,
see https://github.com/hashicorp/hil/issues/50

* update readme

* improve handling CloudFront origin access identity

This allows to still create the CloudFront origin access identity even
if not creating the S3 bucket, and will output the iam_arn of the newly
created CloudFront origin access identity. Addresses issue #7.